Add proptype validation to the Card component
Adds proptype validation to the card component as well as fixing a small
issue with the number inputs and their respective labels. Previously the
input was linked to the label through the use of an id attribute,
however, this caused multiple inputs on the same page to have the same
id. Number inputs are not nested in the label elements.