Add proptype validation to the Item components
Adds proptype validation to the Item component as well as fixes issue
with number input and its respective label. Previously the number input
was linked to the label through an id, however, this caused multiple
number inputs to have the same id in a page. The number input is now
nested inside the label element